Create updated
Dynamic K-factor: The K-factor is adjusted based on the number of games played by a team. Initially, a higher K-factor is used to allow for faster rating adjustments. As the number of games played increases, the K-factor is reduced to slow down rating changes. This accounts for the uncertainty associated with teams with fewer games.

Games played tracking: Each team's rating data now includes the number of games they have played. This information is used to calculate the appropriate K-factor.